backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "fastmcp>=2.12.2",
    "langchain[openai]>=0.3.27",
    "langgraph>=0.6.7",
    "mcp[cli]>=1.13.1",
    "pandas>=2.3.2",
    "pydantic>=2.11.7",
    "pydantic-ai>=1.0.2",
    "python-dotenv>=1.1.1",
    "ruff>=0.12.12",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
